+++ selectors/randomPermutation.h
@@ -7,8 +7,8 @@
 
 class RandomPermutation : public Selector {
 private:
-    vector<int> permutation;
-    int currentId;
+    vector<uint_fast8_t> permutation;
+    uint_fast8_t currentId;
 public:
     explicit RandomPermutation();
     void select(boardType &board) override;
